+++ HISTORY.md
@@ -15,6 +15,7 @@
 
 ### 2025-12-03
 * gui: always normalize SVGs on import
+* gui, objects: prevent dialog boxes to be dropped onto the palette
 
 ### 2025-12-02
 * widgets, gui: prevent a launched tutorial stage from automatically getting keyboard focus